refactor: API 컨트롤러 패키지 구조 개선
employer/worker 패키지 구조를 기능별 패키지 구조로 변경하여 코드 구조를 개선했습니다.
- correctionrequest 패키지로 수정 요청 관련 컨트롤러 통합
- workrecord 패키지로 근무 기록 관련 컨트롤러 통합
