+++ src/main/java/com/example/wagemanager/api/workrecord/WorkerWorkRecordController.java
@@ -1,4 +1,4 @@
-package com.example.wagemanager.api.worker;
+package com.example.wagemanager.api.workrecord;
 
 import com.example.wagemanager.common.dto.ApiResponse;
 import com.example.wagemanager.domain.user.entity.User;
